# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Trace - AI-powered collaborative memory API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Project {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  teamName?: string | null;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export interface ProjectInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  teamName?: string;
}

export interface ProjectUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  teamName?: string;
}

export type ProjectStatsMemoriesByType = {[key: string]: number};

export interface ProjectStats {
  totalMemories: number;
  totalTasks: number;
  openTasks: number;
  memoriesByType: ProjectStatsMemoriesByType;
  recentDecisions?: number;
  completedTasks?: number;
}

export type MemoryType = typeof MemoryType[keyof typeof MemoryType];


export const MemoryType = {
  discussion: 'discussion',
  decision: 'decision',
  task: 'task',
  link: 'link',
  note: 'note',
  screenshot: 'screenshot',
} as const;

export interface Memory {
  id: number;
  projectId: number;
  type: MemoryType;
  content: string;
  /** @nullable */
  summary?: string | null;
  /** @nullable */
  source?: string | null;
  tags?: string[];
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export type MemoryInputType = typeof MemoryInputType[keyof typeof MemoryInputType];


export const MemoryInputType = {
  discussion: 'discussion',
  decision: 'decision',
  task: 'task',
  link: 'link',
  note: 'note',
  screenshot: 'screenshot',
} as const;

export interface MemoryInput {
  type: MemoryInputType;
  /** @minLength 1 */
  content: string;
  source?: string;
  tags?: string[];
}

export interface MemoryUpdate {
  content?: string;
  summary?: string;
  source?: string;
  tags?: string[];
}

export type TaskStatus = typeof TaskStatus[keyof typeof TaskStatus];


export const TaskStatus = {
  open: 'open',
  in_progress: 'in_progress',
  done: 'done',
} as const;

export interface Task {
  id: number;
  projectId: number;
  /** @nullable */
  memoryId?: number | null;
  title: string;
  /** @nullable */
  assignee?: string | null;
  status: TaskStatus;
  /** @nullable */
  dueDate?: string | null;
  createdAt: string;
}

export type TaskInputStatus = typeof TaskInputStatus[keyof typeof TaskInputStatus];


export const TaskInputStatus = {
  open: 'open',
  in_progress: 'in_progress',
  done: 'done',
} as const;

export interface TaskInput {
  /** @minLength 1 */
  title: string;
  assignee?: string;
  status?: TaskInputStatus;
  dueDate?: string;
  memoryId?: number;
}

export type TaskUpdateStatus = typeof TaskUpdateStatus[keyof typeof TaskUpdateStatus];


export const TaskUpdateStatus = {
  open: 'open',
  in_progress: 'in_progress',
  done: 'done',
} as const;

export interface TaskUpdate {
  title?: string;
  assignee?: string;
  status?: TaskUpdateStatus;
  dueDate?: string;
}

export interface QuestionInput {
  /** @minLength 1 */
  question: string;
}

export interface AiAnswer {
  answer: string;
  relevantMemories?: number[];
}

export interface ProjectInsights {
  summary: string;
  keyDecisions: string[];
  actionItems: string[];
  risks: string[];
}

export interface ExtractTasksInput {
  memoryId: number;
}

export type SearchProjectMemoriesParams = {
q: string;
};

export type ListMemoriesParams = {
type?: ListMemoriesType;
};

export type ListMemoriesType = typeof ListMemoriesType[keyof typeof ListMemoriesType];


export const ListMemoriesType = {
  discussion: 'discussion',
  decision: 'decision',
  task: 'task',
  link: 'link',
  note: 'note',
  screenshot: 'screenshot',
} as const;

export type ListTasksParams = {
status?: ListTasksStatus;
};

export type ListTasksStatus = typeof ListTasksStatus[keyof typeof ListTasksStatus];


export const ListTasksStatus = {
  open: 'open',
  in_progress: 'in_progress',
  done: 'done',
} as const;