# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Trace - AI-powered collaborative memory API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}/search:
    get:
      operationId: searchProjectMemories
      tags: [projects]
      summary: Full-text search memories in a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Memory"
  /projects/{projectId}/ask:
    post:
      operationId: askProjectQuestion
      tags: [ai]
      summary: Ask the AI a question about the project's memory
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionInput"
      responses:
        "200":
          description: AI answer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiAnswer"
  /projects/{projectId}/insights:
    get:
      operationId: getProjectInsights
      tags: [ai]
      summary: Get AI-generated insights and action items for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Insights
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectInsights"
  /projects/{projectId}/memories:
    get:
      operationId: listMemories
      tags: [memories]
      summary: List memories for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [discussion, decision, task, link, note, screenshot]
      responses:
        "200":
          description: List of memories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Memory"
    post:
      operationId: createMemory
      tags: [memories]
      summary: Create a new memory entry
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Memory"
  /projects/{projectId}/memories/{memoryId}:
    get:
      operationId: getMemory
      tags: [memories]
      summary: Get a memory entry
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: memoryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Memory found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Memory"
        "404":
          description: Not found
    patch:
      operationId: updateMemory
      tags: [memories]
      summary: Update a memory entry
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: memoryId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Memory"
    delete:
      operationId: deleteMemory
      tags: [memories]
      summary: Delete a memory entry
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: memoryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /projects/{projectId}/memories/{memoryId}/summarize:
    post:
      operationId: summarizeMemory
      tags: [ai]
      summary: AI-summarize a memory entry
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: memoryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Summary generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Memory"
...
components:
  schemas:
...
    Memory:
      type: object
      required: [id, projectId, type, content, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        type:
          type: string
          enum: [discussion, decision, task, link, note, screenshot]
        content:
          type: string
        summary:
          type: ["string", "null"]
        source:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time
    MemoryInput:
      type: object
      required: [type, content]
      properties:
        type:
          type: string
          enum: [discussion, decision, task, link, note, screenshot]
        content:
          type: string
          minLength: 1
        source:
          type: string
        tags:
          type: array
          items:
            type: string
    MemoryUpdate:
      type: object
      properties:
        content:
          type: string
        summary:
          type: string
        source:
          type: string
        tags:
          type: array
          items:
            type: string
...
    QuestionInput:
      type: object
      required: [question]
      properties:
        question:
          type: string
          minLength: 1
    AiAnswer:
      type: object
      required: [answer]
      properties:
        answer:
          type: string
        relevantMemories:
          type: array
          items:
            type: integer
    ProjectInsights:
      type: object
      required: [summary, keyDecisions, actionItems, risks]
      properties:
        summary:
          type: string
        keyDecisions:
          type: array
          items:
            type: string
        actionItems:
          type: array
          items:
            type: string
        risks:
          type: array
          items:
            type: string
...