# lib/api-zod/src/generated/types/questionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Trace - AI-powered collaborative memory API
 * OpenAPI spec version: 0.1.0
 */

export interface QuestionInput {
  /** @minLength 1 */
  question: string;
}